MM-36766 - Remove RHS tabs (#627)
* remove rhs tabs

* e2e tests

* simplify selector
