+++ tests-e2e/cypress/integration/frontstage/playbook_run_rhs_checklist_spec.js
@@ -121,9 +121,6 @@ describe('playbook run rhs checklist', () => {
             // * Verify the playbook run RHS is open.
             cy.get('#rhsContainer').should('exist').within(() => {
                 cy.findByText(playbookRunName).should('exist');
-
-                // # Select the checklists tab
-                cy.findByTestId('checklists').click();
             });
         });
 
@@ -164,9 +161,6 @@ describe('playbook run rhs checklist', () => {
             cy.visit('/ad-1/channels/' + playbookRunChannelName);
 
             cy.get('#rhsContainer').should('exist').within(() => {
-                // # Select the checklists tab
-                cy.findByTestId('checklists').click();
-
                 // * Verify the invalid command still has not yet been run.
                 cy.findAllByTestId('run').eq(0).should('have.text', 'Run');
